requirements.txt
requests
pyTelegramBotAPI
google-genai
python-dotenv
APScheduler==3.10.4
pytz==2024.1